+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Improved drawing loop by using hardware acceleration with vsync
 - Modified `draw_loop` in the `WidgetCache` to return true/false for display swap/refresh
+- Modified `BaseWidget` to use `Texture` for drawing (Part of work for #29)
+- Modified `CanvasWidget` to use `Texture` for drawing (Part of work for #29)
 
 ## 0.4.22